Extract order book into separate class
- Add some simple logs.
